Separate core library from renderer + remove some apps
