+++ apps/firework-explosion.ts
@@ -1,13 +1,13 @@
 import * as PIXI from "pixi.js";
-import { ParticleSystem } from "particleSystem";
-import { PointGenerator } from "generators/pointGenerator";
-import { LifeTimeDestructor } from "destructors/lifeTimeDestructor";
-import { LifeTimeRange } from "initializers/lifeTimeRange";
+import { ParticleSystem } from "core/particleSystem";
+import { PointGenerator } from "core/generators/pointGenerator";
+import { LifeTimeDestructor } from "core/destructors/lifeTimeDestructor";
+import { LifeTimeRange } from "core/initializers/lifeTimeRange";
 import { Renderer } from "renderer/renderer";
-import { RandomAngleVelocity } from "initializers/randomAngleVelocity";
-import { DeaccelerationOverLifetime } from "modifiers/deaccelerationOverLifetime";
-import { AlphaOverLifetime } from "modifiers/alphaOverLifetime";
-import { EasingFunctions } from "easing";
+import { RandomAngleVelocity } from "core/initializers/randomAngleVelocity";
+import { DeaccelerationOverLifetime } from "core/modifiers/deaccelerationOverLifetime";
+import { AlphaOverLifetime } from "core/modifiers/alphaOverLifetime";
+import { EasingFunctions } from "core/easing";
 
 document.body.style.margin = "0px 0px";
 document.body.style.width = "100vw";
