+++ src/core/destructors/alphaDestructor.ts
