+++ src/core/easing.ts
