+++ src/core/generators/circleExteriorGenerator.ts
