+++ src/core/generators/pointGenerator.ts
