+++ src/core/module.ts
