+++ src/core/particle.ts
