+++ src/core/particleSystem.ts
