+++ src/core/shapes/shapes.ts
@@ -1,4 +1,4 @@
-import { Position } from "types";
+import { Position } from "core/types";
 import { Triangle } from "./triangle";
 
 /**
