+++ src/core/utilities.ts
@@ -1,5 +1,4 @@
-import { Color } from "types";
-import { Position } from "types";
+import { Color, Position } from "core/types";
 
 export const lerp = (x: number, y: number, a: number) => x * (1 - a) + y * a;
 export const clamp = (a: number, min = 0, max = 1) => Math.min(max, Math.max(min, a));
